requirements.txt
openai==0.28.0
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
tqdm==4.66.1
flask==2.3.3
waitress==2.1.2
pytest==7.4.0
pytest-cov==4.1.0 